htlcswitch: add LinkError interface
This commit adds a LinkError interface which
will be used to wrap wire messages within the
switch. This interface is intended to place a
barrier between the switch and the wire, as
well as providing the ability to wrap internal
errors with additional metadata without risk
of leaking additional information onto the wire.