htlcswitch: add notifications for forwards
This commit adds notifications for htlcs which are
forwarded through our node. Forwards are notified
when the htlc is added on our ougoing link, settles
when we send a settle message to the downstream peer.
If a failure occurs, we check whether it occurred
at our node, then notify a link or forwarding
failure accordingly.

Note that this change also adds forward event
notifications for sends which are initiated
by our node because the handling code for adding
a htlc which originates from our node is the same
as that for handling forwards. Subsequent settles
or failures are not noitfied for local sends in
this commit because they are handled separately
(so will be added in the following commit).